// AutoVideoMetaLocalize/client/generated-sources/openapi/models/Video.ts
export interface Video {
    /**
     * 
     * @type {VideoAgeGating}
     * @memberof Video
     */
    ageGating?: VideoAgeGating;
    /**
     * 
     * @type {VideoContentDetails}
     * @memberof Video
     */
    contentDetails?: VideoContentDetails;
    /**
     * 
     * @type {string}
     * @memberof Video
     */
    eTag?: string | null;
    /**
     * 
     * @type {VideoFileDetails}
     * @memberof Video
     */
    fileDetails?: VideoFileDetails;
    /**
     * 
     * @type {string}
     * @memberof Video
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Video
     */
    kind?: string | null;
    /**
     * 
     * @type {VideoLiveStreamingDetails}
     * @memberof Video
     */
    liveStreamingDetails?: VideoLiveStreamingDetails;
    /**
     * 
     * @type {{ [key: string]: VideoLocalization; }}
     * @memberof Video
     */
    localizations?: { [key: string]: VideoLocalization; } | null;
    /**
     * 
     * @type {VideoMonetizationDetails}
     * @memberof Video
     */
    monetizationDetails?: VideoMonetizationDetails;
    /**
     * 
     * @type {VideoPlayer}
     * @memberof Video
     */
    player?: VideoPlayer;
    /**
     * 
     * @type {VideoProcessingDetails}
     * @memberof Video
     */
    processingDetails?: VideoProcessingDetails;
    /**
     * 
     * @type {VideoProjectDetails}
     * @memberof Video
     */
    projectDetails?: VideoProjectDetails;
    /**
     * 
     * @type {VideoRecordingDetails}
     * @memberof Video
     */
    recordingDetails?: VideoRecordingDetails;
    /**
     * 
     * @type {VideoSnippet}
     * @memberof Video
     */
    snippet?: VideoSnippet;
    /**
     * 
     * @type {VideoStatistics}
     * @memberof Video
     */
    statistics?: VideoStatistics;
    /**
     * 
     * @type {VideoStatus}
     * @memberof Video
     */
    status?: VideoStatus;
    /**
     * 
     * @type {VideoSuggestions}
     * @memberof Video
     */
    suggestions?: VideoSuggestions;
    /**
     * 
     * @type {VideoTopicDetails}
     * @memberof Video
     */
    topicDetails?: VideoTopicDetails;
}