// AutoVideoMetaLocalize/client/generated-sources/openapi/models/VideoLocalization.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AutoVideoMetaLocalize
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface VideoLocalization
 */
export interface VideoLocalization {
    /**
     * 
     * @type {string}
     * @memberof VideoLocalization
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof VideoLocalization
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof VideoLocalization
     */
    eTag?: string | null;
}

export function VideoLocalizationFromJSON(json: any): VideoLocalization {
    return VideoLocalizationFromJSONTyped(json, false);
}

export function VideoLocalizationFromJSONTyped(json: any, ignoreDiscriminator: boolean): VideoLocalization {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'description': !exists(json, 'description') ? undefined : json['description'],
        'title': !exists(json, 'title') ? undefined : json['title'],
        'eTag': !exists(json, 'eTag') ? undefined : json['eTag'],
    };
}

export function VideoLocalizationToJSON(value?: VideoLocalization | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'description': value.description,
        'title': value.title,
        'eTag': value.eTag,
    };
}